[Chore] #17 - 파일 위치 수정
