+++ lib/application/tabBar/model/tab_model.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:money_planet/presentaion/register/view/login_screen.dart';
 
 import '../../../presentaion/chart/view/chart_screen.dart';
 import '../../../presentaion/diary/view/diary_screen.dart';
@@ -13,7 +12,7 @@ class TabModel {
   final List<Widget> pages = [
     const HomeScreen(),
     const DiaryScreen(),
-    const LoginScreen(),
+    const RegisterScreen(),
     const ChartScreen(),
     const MyPageScreen(),
   ];
