+++ lib/presentaion/onboarding/view/selecttype_screen.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class SelectTypeScreen1 extends StatelessWidget{
+  const SelectTypeScreen1({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+    backgroundColor: Color(0xFFD7E3FF),
+        body: SafeArea(
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+
+        ],
+    ),
+    ),
+    );
+  }
+
+}
