+++ lib/presentaion/onboarding/view/signup_screen.dart
@@ -1,15 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:money_planet/global/theme/colors.dart';
+import 'package:money_planet/global/theme/textStyles.dart';
+import 'package:money_planet/presentaion/onboarding/view/signup_screen2.dart';
 
-import '../../../global/theme/textStyles.dart';
-import 'guidetype_screen.dart';
 import 'login_screen.dart';
 
-class RegisterScreen2 extends StatelessWidget {
-  const RegisterScreen2({super.key});
+class SignUpScreen extends StatelessWidget {
+  const SignUpScreen({super.key});
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context) {// 배경
+    final buttonColor = const Color(0xFF9292A1); // 회색 버튼
 
     return Scaffold(
       backgroundColor: neutral_900,
@@ -17,10 +18,12 @@ class RegisterScreen2 extends StatelessWidget {
         child: Column(
           children: [
             const SizedBox(height: 100),
-            Center(
-              child: Text(
-                '회원 가입',
-                style: customTextStyle(fontFamily: Pretendard_Bold_24,color: secondary_050_1)
+            Container(
+              child: Center(
+                child:  Text(
+                  '회원 가입',
+                  style: customTextStyle(fontFamily: Pretendard_Bold_24,color: secondary_050_1)
+                ),
               ),
             ),
             const SizedBox(height: 60),
@@ -39,12 +42,12 @@ class RegisterScreen2 extends StatelessWidget {
                   children: [
                     Align(
                       alignment: Alignment.centerLeft,
-                      child: Text('비밀번호', style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
+                      child: Text('이름', style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
                     ),
                     const SizedBox(height: 6),
                     TextField(
                       decoration: InputDecoration(
-                        hintText: '비밀번호를 입력해주세요',
+                        hintText: '이름을 입력해주세요',
                         filled: true,
                         fillColor: Colors.white,
                         border: OutlineInputBorder(
@@ -60,13 +63,13 @@ class RegisterScreen2 extends StatelessWidget {
                     const SizedBox(height: 16),
                     Align(
                       alignment: Alignment.centerLeft,
-                      child: Text('비밀번호 확인',
+                      child: Text('이메일',
                           style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
                     ),
                     const SizedBox(height: 6),
                     TextField(
                       decoration: InputDecoration(
-                        hintText: '비밀번호를 다시 입력해 주세요',
+                        hintText: 'example@example.com',
                         filled: true,
                         fillColor: Colors.white,
                         border: OutlineInputBorder(
@@ -79,7 +82,51 @@ class RegisterScreen2 extends StatelessWidget {
                         ),
                       ),
                     ),
-                    const SizedBox(height: 240),
+                    const SizedBox(height: 16),
+                    Align(
+                      alignment: Alignment.centerLeft,
+                      child: Text('전화번호',
+                          style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
+                    ),
+                    const SizedBox(height: 6),
+                    TextField(
+                      decoration: InputDecoration(
+                        hintText: '+123 456 789',
+                        filled: true,
+                        fillColor: Colors.white,
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(12),
+                          borderSide: BorderSide.none,
+                        ),
+                        contentPadding: const EdgeInsets.symmetric(
+                          horizontal: 16,
+                          vertical: 12,
+                        ),
+                      ),
+                    ),
+                    const SizedBox(height: 16),
+                    Align(
+                      alignment: Alignment.centerLeft,
+                      child: Text('생일',
+                          style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
+                    ),
+                    const SizedBox(height: 6),
+                    TextField(
+                      decoration: InputDecoration(
+                        hintText: 'DD/MM/YY',
+                        filled: true,
+                        fillColor: Colors.white,
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(12),
+                          borderSide: BorderSide.none,
+                        ),
+                        contentPadding: const EdgeInsets.symmetric(
+                          horizontal: 16,
+                          vertical: 12,
+                        ),
+                      ),
+                    ),
+                    const SizedBox(height: 40),
                     SizedBox(
                       width: double.infinity,
                       height: 56,
@@ -87,17 +134,17 @@ class RegisterScreen2 extends StatelessWidget {
                         onPressed: () {
                           Navigator.push(
                             context,
-                            MaterialPageRoute(builder: (context) => const GuideTypeScreen()),
+                            MaterialPageRoute(builder: (context) => const SignUpScreen2()),
                           );
                         },
                         style: ElevatedButton.styleFrom(
-                          backgroundColor: neutral_300,
+                          backgroundColor: buttonColor,
                           shape: RoundedRectangleBorder(
                             borderRadius: BorderRadius.circular(12),
                           ),
                         ),
                         child: Text('다음',
-                            style: customTextStyle(fontFamily: Pretendard_Medium_18,color: Colors.white)),
+                          style: customTextStyle(fontFamily: Pretendard_Medium_18,color: Colors.white)),
                       ),
                     ),
                     const SizedBox(height: 16),
