+++ lib/presentaion/register/view/register_screen.dart
@@ -1,172 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:money_planet/global/theme/colors.dart';
-import 'package:money_planet/global/theme/textStyles.dart';
-import 'package:money_planet/presentaion/register/view/register_screen2.dart';
-
-import 'login_screen.dart';
 
 class RegisterScreen extends StatelessWidget {
   const RegisterScreen({super.key});
 
   @override
-  Widget build(BuildContext context) {// 배경
-    final buttonColor = const Color(0xFF9292A1); // 회색 버튼
-
-    return Scaffold(
-      backgroundColor: neutral_900,
-      body: SafeArea(
-        child: Column(
-          children: [
-            const SizedBox(height: 100),
-            Container(
-              child: Center(
-                child:  Text(
-                  '회원 가입',
-                  style: customTextStyle(fontFamily: Pretendard_Bold_24,color: secondary_050_1)
-                ),
-              ),
-            ),
-            const SizedBox(height: 60),
-            Expanded(
-              child: Container(
-                padding: const EdgeInsets.symmetric(horizontal: 40),
-                decoration: BoxDecoration(
-                  color: primary_050,
-                  borderRadius: const BorderRadius.only(
-                    topLeft: Radius.circular(24),
-                    topRight: Radius.circular(24),
-                  ),
-                ),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Align(
-                      alignment: Alignment.centerLeft,
-                      child: Text('이름', style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
-                    ),
-                    const SizedBox(height: 6),
-                    TextField(
-                      decoration: InputDecoration(
-                        hintText: '이름을 입력해주세요',
-                        filled: true,
-                        fillColor: Colors.white,
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                          borderSide: BorderSide.none,
-                        ),
-                        contentPadding: const EdgeInsets.symmetric(
-                          horizontal: 16,
-                          vertical: 12,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Align(
-                      alignment: Alignment.centerLeft,
-                      child: Text('이메일',
-                          style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
-                    ),
-                    const SizedBox(height: 6),
-                    TextField(
-                      decoration: InputDecoration(
-                        hintText: 'example@example.com',
-                        filled: true,
-                        fillColor: Colors.white,
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                          borderSide: BorderSide.none,
-                        ),
-                        contentPadding: const EdgeInsets.symmetric(
-                          horizontal: 16,
-                          vertical: 12,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Align(
-                      alignment: Alignment.centerLeft,
-                      child: Text('전화번호',
-                          style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
-                    ),
-                    const SizedBox(height: 6),
-                    TextField(
-                      decoration: InputDecoration(
-                        hintText: '+123 456 789',
-                        filled: true,
-                        fillColor: Colors.white,
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                          borderSide: BorderSide.none,
-                        ),
-                        contentPadding: const EdgeInsets.symmetric(
-                          horizontal: 16,
-                          vertical: 12,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Align(
-                      alignment: Alignment.centerLeft,
-                      child: Text('생일',
-                          style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black)),
-                    ),
-                    const SizedBox(height: 6),
-                    TextField(
-                      decoration: InputDecoration(
-                        hintText: 'DD/MM/YY',
-                        filled: true,
-                        fillColor: Colors.white,
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                          borderSide: BorderSide.none,
-                        ),
-                        contentPadding: const EdgeInsets.symmetric(
-                          horizontal: 16,
-                          vertical: 12,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(height: 40),
-                    SizedBox(
-                      width: double.infinity,
-                      height: 56,
-                      child: ElevatedButton(
-                        onPressed: () {
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(builder: (context) => const RegisterScreen2()),
-                          );
-                        },
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: buttonColor,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(12),
-                          ),
-                        ),
-                        child: Text('다음',
-                          style: customTextStyle(fontFamily: Pretendard_Medium_18,color: Colors.white)),
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    GestureDetector(
-                      onTap: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(builder: (context) => const LoginScreen()),
-                        );
-                      },
-                      child: Text(
-                        '이미 계정이 있으신가요? 로그인하기',
-                        style: customTextStyle(fontFamily: Pretendard_Medium_14,color: Colors.black),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
+  Widget build(BuildContext context) {
+    return const Center(child: Text('등록 화면'));
   }
 }
